# Run golangci-lint from docker-compose