--- a/geom/alg_set_op_test.go
+++ b/geom/alg_set_op_test.go
@@ -1294,7 +1294,6 @@ func TestBinaryOpBothInputsEmpty(t *testing.T) {
 					t.Errorf("got=%v but want=FFFFFFFF2", got)
 				}
 			})
-
 		})
 	}
 }