--- a/geom/dcel_input.go
+++ b/geom/dcel_input.go
@@ -109,7 +109,6 @@ func (d *doublyConnectedEdgeList) addLineString(ls LineString, operand operand, 
 				}
 			}
 		}
-
 	})
 }