--- a/geom/twkb_export_test.go
+++ b/geom/twkb_export_test.go
@@ -1,4 +1,6 @@
 package geom
 
-var DecodeZigZagInt64 = decodeZigZagInt64
-var EncodeZigZagInt64 = encodeZigZagInt64
+var (
+	DecodeZigZagInt64 = decodeZigZagInt64
+	EncodeZigZagInt64 = encodeZigZagInt64
+)