## Description

Please describe at a high level what the change is and why it is useful or
required.

## Check List

Have you:

- Added unit tests?

- Add cmprefimpl tests? (if appropriate?)

- Updated release notes? (if appropriate?)

## Related Issue

- Please link to the related issue(s).

## Benchmark Results

- Please paste benchmark results here. The benchmarks can be run using the
  `run_benchmarks.sh` script.

<details>

<summary>Click to expand</summary>

```
COMPARISON
name                                                         old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:amd64
LineEnvelope/0-12                                              2.42ns ± 7%    2.45ns ± 6%     ~     (p=0.433 n=13+15)
LineEnvelope/1-12                                              2.64ns ±37%    2.41ns ± 3%     ~     (p=0.102 n=14+15)
LineEnvelope/2-12                                              2.21ns ± 5%    2.16ns ± 1%   -1.91%  (p=0.024 n=14+12)
LineEnvelope/3-12                                              2.23ns ±18%    2.17ns ± 7%     ~     (p=0.676 n=14+13)
MarshalWKB/polygon/n=10-12                                      189ns ±14%     182ns ±15%     ~     (p=0.190 n=15+14)
MarshalWKB/polygon/n=100-12                                     521ns ±16%     545ns ±24%     ~     (p=0.497 n=14+15)
MarshalWKB/polygon/n=1000-12                                   4.15µs ±29%    3.93µs ±28%     ~     (p=0.389 n=15+15)
MarshalWKB/polygon/n=10000-12                                  38.0µs ±19%    37.7µs ±18%     ~     (p=0.981 n=13+14)
UnmarshalWKB/polygon/n=10-12                                    354ns ±26%     331ns ±15%     ~     (p=0.217 n=15+14)
UnmarshalWKB/polygon/n=100-12                                   693ns ±19%     632ns ± 9%   -8.78%  (p=0.019 n=15+13)
UnmarshalWKB/polygon/n=1000-12                                 4.41µs ±23%    4.06µs ±16%     ~     (p=0.051 n=15+14)
UnmarshalWKB/polygon/n=10000-12                                39.9µs ±12%    42.9µs ±28%     ~     (p=0.201 n=13+15)
IntersectsLineStringWithLineString/n=10-12                     1.56µs ± 9%    1.52µs ±13%     ~     (p=0.437 n=13+14)
IntersectsLineStringWithLineString/n=100-12                    21.4µs ±23%    20.9µs ±17%     ~     (p=0.427 n=14+14)
IntersectsLineStringWithLineString/n=1000-12                    204µs ±15%     191µs ±10%   -6.13%  (p=0.025 n=15+13)
IntersectsLineStringWithLineString/n=10000-12                  3.35ms ±19%    3.12ms ± 9%     ~     (p=0.123 n=15+14)
IntersectsMultiPointWithMultiPoint/n=20-12                      872ns ±24%     811ns ± 4%   -6.98%  (p=0.008 n=15+13)
IntersectsMultiPointWithMultiPoint/n=200-12                    10.2µs ±24%     9.6µs ± 4%     ~     (p=0.650 n=15+13)
IntersectsMultiPointWithMultiPoint/n=2000-12                    103µs ±18%      96µs ±11%     ~     (p=0.061 n=14+13)
IntersectsMultiPointWithMultiPoint/n=20000-12                  1.20ms ±15%    1.19ms ±21%     ~     (p=0.653 n=15+15)
PolygonSingleRingValidation/n=10-12                            2.57µs ±30%    2.51µs ±10%     ~     (p=0.839 n=14+14)
PolygonSingleRingValidation/n=100-12                           35.1µs ±37%    33.1µs ±11%     ~     (p=0.511 n=14+14)
PolygonSingleRingValidation/n=1000-12                           414µs ±24%     412µs ±15%     ~     (p=0.734 n=14+14)
PolygonSingleRingValidation/n=10000-12                         5.51ms ±21%    5.39ms ±18%     ~     (p=0.561 n=14+15)
PolygonMultipleRingsValidation/n=4-12                          7.19µs ±12%    7.13µs ±18%     ~     (p=0.586 n=13+15)
PolygonMultipleRingsValidation/n=36-12                         60.7µs ±16%    60.6µs ±22%     ~     (p=0.683 n=13+15)
PolygonMultipleRingsValidation/n=400-12                         805µs ±15%     808µs ±14%     ~     (p=1.000 n=14+15)
PolygonMultipleRingsValidation/n=4096-12                       9.74ms ±19%    9.80ms ±24%     ~     (p=0.838 n=15+15)
PolygonZigZagRingsValidation/n=10-12                           11.9µs ±16%    11.4µs ±11%     ~     (p=0.169 n=14+13)
PolygonZigZagRingsValidation/n=100-12                           143µs ±35%     135µs ±13%     ~     (p=0.227 n=14+14)
PolygonZigZagRingsValidation/n=1000-12                         1.55ms ±19%    1.56ms ±23%     ~     (p=0.467 n=13+15)
PolygonZigZagRingsValidation/n=10000-12                        20.8ms ±12%    20.2ms ±25%     ~     (p=0.116 n=13+14)
PolygonAnnulusValidation/n=10-12                               4.19µs ±17%    4.06µs ±16%     ~     (p=0.387 n=13+15)
PolygonAnnulusValidation/n=100-12                              37.7µs ±31%    34.9µs ±17%     ~     (p=0.270 n=14+15)
PolygonAnnulusValidation/n=1000-12                              608µs ±26%     613µs ±21%     ~     (p=0.618 n=13+15)
PolygonAnnulusValidation/n=10000-12                            7.12ms ±11%    7.40ms ±15%     ~     (p=0.194 n=14+14)
MultipolygonValidation/n=1-12                                   457ns ±14%     477ns ±32%     ~     (p=0.462 n=14+14)
MultipolygonValidation/n=4-12                                  1.04µs ±29%    1.02µs ±19%     ~     (p=0.910 n=14+14)
MultipolygonValidation/n=16-12                                 4.00µs ±15%    4.16µs ±18%     ~     (p=0.359 n=14+15)
MultipolygonValidation/n=64-12                                 18.4µs ±17%    19.5µs ±19%     ~     (p=0.164 n=14+14)
MultipolygonValidation/n=256-12                                 111µs ±12%     118µs ±25%     ~     (p=0.329 n=14+14)
MultipolygonValidation/n=1024-12                                528µs ±13%     578µs ±30%     ~     (p=0.070 n=15+14)
MultiPolygonTwoCircles/n=10-12                                 4.01µs ±20%    4.13µs ±13%     ~     (p=0.425 n=15+14)
MultiPolygonTwoCircles/n=100-12                                42.5µs ±14%    46.6µs ±36%     ~     (p=0.146 n=14+15)
MultiPolygonTwoCircles/n=1000-12                                455µs ±19%     455µs ±15%     ~     (p=1.000 n=15+14)
MultiPolygonTwoCircles/n=10000-12                              7.27ms ±19%    7.26ms ±12%     ~     (p=1.000 n=15+13)
MultiPolygonMultipleTouchingPoints/n=1-12                      5.42µs ±18%    6.38µs ±71%     ~     (p=0.561 n=15+14)
MultiPolygonMultipleTouchingPoints/n=10-12                     39.6µs ±12%    48.8µs ±60%  +23.28%  (p=0.008 n=15+14)
MultiPolygonMultipleTouchingPoints/n=100-12                     480µs ±12%     516µs ±18%     ~     (p=0.134 n=15+14)
MultiPolygonMultipleTouchingPoints/n=1000-12                   5.58ms ±11%    6.62ms ±57%  +18.80%  (p=0.024 n=14+14)
WKTParsing/point-12                                            1.90µs ±14%    2.23µs ±84%     ~     (p=0.133 n=15+13)
DistancePolygonToPolygonOrdering/n=100_swap=false-12           53.5µs ±25%    64.6µs ±76%     ~     (p=0.146 n=15+14)
DistancePolygonToPolygonOrdering/n=100_swap=true-12            51.3µs ± 9%    62.4µs ±62%  +21.58%  (p=0.039 n=14+14)
DistancePolygonToPolygonOrdering/n=1000_swap=false-12           764µs ±25%     832µs ±28%   +8.93%  (p=0.031 n=14+14)
DistancePolygonToPolygonOrdering/n=1000_swap=true-12            763µs ±19%     904µs ±47%  +18.53%  (p=0.009 n=14+15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-12     5.47µs ±17%    5.94µs ±13%   +8.57%  (p=0.020 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-12      5.76µs ±11%    6.04µs ±24%     ~     (p=0.408 n=14+14)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-12    61.9µs ±17%    63.1µs ±15%     ~     (p=0.297 n=14+12)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-12     60.0µs ±14%    65.0µs ±36%     ~     (p=0.158 n=15+14)
MultiLineStringIsSimpleManyLineStrings/n=100-12                53.8µs ±19%    59.2µs ±44%     ~     (p=0.259 n=13+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-12                575µs ±15%     623µs ±18%   +8.35%  (p=0.029 n=14+13)
ForceCWandForceCCW/0-12                                        32.2ns ±15%    31.6ns ± 7%     ~     (p=0.529 n=12+13)
ForceCWandForceCCW/0#01-12                                      230ns ±21%     230ns ±16%     ~     (p=0.840 n=11+14)
ForceCWandForceCCW/1-12                                       84.8ns ±201%    34.1ns ±23%     ~     (p=0.924 n=15+14)
ForceCWandForceCCW/1#01-12                                      175ns ±82%     239ns ±20%  +36.34%  (p=0.018 n=15+14)
ForceCWandForceCCW/2-12                                        58.4ns ± 7%    59.9ns ±17%     ~     (p=0.440 n=13+13)
ForceCWandForceCCW/2#01-12                                      379ns ±21%     375ns ±11%     ~     (p=1.000 n=14+12)
ForceCWandForceCCW/3-12                                        57.7ns ± 2%    58.5ns ±11%     ~     (p=0.832 n=12+12)
ForceCWandForceCCW/3#01-12                                      372ns ±19%     395ns ±27%     ~     (p=0.347 n=12+13)
ForceCWandForceCCW/4-12                                        88.6ns ±15%    87.7ns ±15%     ~     (p=0.496 n=14+13)
ForceCWandForceCCW/4#01-12                                      600ns ±13%     640ns ±27%     ~     (p=0.125 n=14+14)
ForceCWandForceCCW/5-12                                        84.8ns ± 8%    87.8ns ±28%     ~     (p=0.470 n=12+13)
ForceCWandForceCCW/5#01-12                                      584ns ±18%     608ns ±17%     ~     (p=0.223 n=13+13)
ForceCWandForceCCW/6-12                                         121ns ±11%     123ns ±12%   +1.62%  (p=0.046 n=12+12)
ForceCWandForceCCW/6#01-12                                      886ns ±13%     914ns ±18%     ~     (p=0.347 n=12+12)
EnvelopeTransformXY-12                                          280ns ± 6%     279ns ± 6%     ~     (p=0.804 n=14+14)
WKBParse/0-12                                                   771ns ±13%     789ns ±18%     ~     (p=0.870 n=15+15)
WKBParse/1-12                                                   770ns ±12%     777ns ±16%     ~     (p=0.983 n=15+14)
WKBParse/2-12                                                   829ns ± 4%     895ns ±23%     ~     (p=0.144 n=13+14)
WKBParse/3-12                                                   835ns ± 3%     892ns ±13%   +6.89%  (p=0.006 n=12+15)
WKBParse/4-12                                                   976ns ±11%    1051ns ±21%   +7.73%  (p=0.004 n=13+15)
WKBParse/5-12                                                   763ns ± 5%     799ns ±11%   +4.74%  (p=0.010 n=13+15)
WKBParse/6-12                                                   850ns ± 8%     879ns ±11%     ~     (p=0.254 n=13+15)
WKBParse/7-12                                                   849ns ± 8%     922ns ±27%     ~     (p=0.080 n=13+15)
WKBParse/8-12                                                   991ns ± 3%    1063ns ±16%   +7.26%  (p=0.004 n=14+14)
WKBParse/9-12                                                   636ns ±19%     644ns ±24%     ~     (p=0.691 n=15+15)
WKBParse/10-12                                                  611ns ± 6%     619ns ± 8%     ~     (p=0.579 n=14+14)
WKBParse/11-12                                                  619ns ±12%     621ns ± 7%     ~     (p=0.839 n=14+14)
WKBParse/12-12                                                  598ns ± 2%     617ns ± 8%   +3.14%  (p=0.037 n=13+15)
WKBParse/13-12                                                 1.04µs ± 2%    1.10µs ±17%   +5.44%  (p=0.014 n=12+15)
WKBParse/14-12                                                 1.19µs ± 2%    1.24µs ± 7%   +3.78%  (p=0.000 n=13+14)
WKBParse/15-12                                                 1.19µs ± 3%    1.27µs ±13%   +6.63%  (p=0.012 n=13+15)
WKBParse/16-12                                                 1.41µs ± 5%    1.48µs ±13%     ~     (p=0.063 n=13+15)
WKBParse/17-12                                                 1.19µs ± 2%    1.23µs ±12%   +3.55%  (p=0.016 n=13+14)
WKBParse/18-12                                                 1.63µs ±11%    1.59µs ± 9%     ~     (p=0.383 n=15+14)
WKBParse/19-12                                                 1.62µs ±10%    1.69µs ±27%     ~     (p=0.911 n=15+15)
WKBParse/20-12                                                 1.88µs ±10%    1.94µs ±16%     ~     (p=0.454 n=15+15)
WKBParse/21-12                                                  582ns ± 4%     580ns ± 3%     ~     (p=0.631 n=14+12)
WKBParse/22-12                                                  581ns ± 3%     585ns ± 7%     ~     (p=0.790 n=13+13)
WKBParse/23-12                                                  577ns ± 4%     581ns ± 4%     ~     (p=0.395 n=13+14)
WKBParse/24-12                                                  572ns ± 4%     625ns ±23%   +9.34%  (p=0.001 n=14+15)
WKBParse/25-12                                                 2.62µs ± 8%    2.68µs ± 8%     ~     (p=0.169 n=13+13)
WKBParse/26-12                                                 3.23µs ± 6%    3.25µs ± 7%     ~     (p=0.441 n=13+13)
WKBParse/27-12                                                 3.43µs ±14%    3.35µs ±16%     ~     (p=0.174 n=14+14)
WKBParse/28-12                                                 4.24µs ±11%    4.44µs ±12%     ~     (p=0.075 n=14+14)
WKBParse/29-12                                                  575ns ± 3%     630ns ±17%   +9.52%  (p=0.000 n=12+15)
WKBParse/30-12                                                  601ns ± 8%     599ns ±11%     ~     (p=0.974 n=15+13)
WKBParse/31-12                                                  580ns ± 3%     600ns ±11%     ~     (p=0.564 n=12+13)
WKBParse/32-12                                                  573ns ± 3%     596ns ±10%     ~     (p=0.100 n=13+12)
WKBParse/33-12                                                 1.01µs ± 5%    1.09µs ±16%   +7.68%  (p=0.001 n=15+12)
WKBParse/34-12                                                 1.20µs ±15%    1.25µs ±14%     ~     (p=0.152 n=15+13)
WKBParse/35-12                                                 1.19µs ±10%    1.33µs ±16%  +12.36%  (p=0.008 n=15+15)
WKBParse/36-12                                                 1.32µs ±16%    1.31µs ±21%     ~     (p=0.991 n=15+13)
WKBParse/37-12                                                 1.51µs ±19%    1.49µs ±14%     ~     (p=1.000 n=15+13)
WKBParse/38-12                                                 1.76µs ±13%    1.80µs ±19%     ~     (p=0.505 n=15+14)
WKBParse/39-12                                                 1.70µs ± 8%    1.80µs ±20%     ~     (p=0.135 n=13+13)
WKBParse/40-12                                                 1.91µs ± 7%    2.00µs ±24%     ~     (p=0.315 n=15+14)
WKBParse/41-12                                                  598ns ±10%     617ns ±27%     ~     (p=0.496 n=15+13)
WKBParse/42-12                                                  580ns ± 2%     597ns ±11%     ~     (p=0.381 n=12+14)
WKBParse/43-12                                                  600ns ±13%     605ns ±19%     ~     (p=0.955 n=14+14)
WKBParse/44-12                                                  602ns ±16%     601ns ±15%     ~     (p=0.880 n=15+14)
WKBParse/45-12                                                 1.62µs ± 7%    1.64µs ± 6%     ~     (p=0.735 n=13+13)
WKBParse/46-12                                                 1.93µs ±12%    1.93µs ±11%     ~     (p=0.805 n=15+14)
WKBParse/47-12                                                 1.85µs ± 6%    1.94µs ±16%     ~     (p=0.382 n=15+14)
WKBParse/48-12                                                 2.07µs ± 8%    2.21µs ±23%     ~     (p=0.095 n=15+15)
WKBParse/49-12                                                 2.21µs ± 9%    2.28µs ±22%     ~     (p=0.169 n=14+13)
WKBParse/50-12                                                 2.71µs ± 4%    2.80µs ±12%     ~     (p=0.306 n=14+14)
WKBParse/51-12                                                 2.74µs ±11%    2.82µs ±13%     ~     (p=0.540 n=14+15)
WKBParse/52-12                                                 3.22µs ±14%    3.12µs ± 4%     ~     (p=0.467 n=15+13)
WKBParse/53-12                                                  593ns ± 7%     580ns ± 2%     ~     (p=0.280 n=14+13)
WKBParse/54-12                                                  607ns ±13%     601ns ± 7%     ~     (p=0.846 n=15+15)
WKBParse/55-12                                                  596ns ± 9%     646ns ±34%     ~     (p=0.116 n=15+15)
WKBParse/56-12                                                  580ns ± 9%     585ns ± 7%     ~     (p=0.111 n=14+14)
WKBParse/57-12                                                 3.35µs ± 9%    3.50µs ±14%     ~     (p=0.451 n=14+15)
WKBParse/58-12                                                 3.94µs ± 4%    4.11µs ±12%     ~     (p=0.066 n=13+14)
WKBParse/59-12                                                 4.12µs ±12%    4.25µs ±25%     ~     (p=0.567 n=15+15)
WKBParse/60-12                                                 4.69µs ± 4%    5.08µs ±27%     ~     (p=0.075 n=12+15)
WKBParse/61-12                                                  585ns ± 4%     586ns ± 8%     ~     (p=0.858 n=13+14)
WKBParse/62-12                                                  611ns ±12%     587ns ± 4%     ~     (p=0.409 n=14+13)
WKBParse/63-12                                                  575ns ± 3%     607ns ±19%     ~     (p=0.129 n=12+15)
WKBParse/64-12                                                  586ns ±10%     595ns ±17%     ~     (p=0.992 n=14+15)
WKBParse/65-12                                                 1.67µs ±15%    1.62µs ± 7%     ~     (p=0.285 n=14+14)
WKBParse/66-12                                                 1.95µs ±21%    1.88µs ±15%     ~     (p=0.720 n=14+13)
WKBParse/67-12                                                 1.96µs ±28%    1.86µs ±10%     ~     (p=0.756 n=15+14)
WKBParse/68-12                                                 2.10µs ±26%    1.98µs ± 7%     ~     (p=0.315 n=15+14)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:amd64
IntersectionWithoutValidation/n=10-12                          29.8µs ±25%    27.0µs ± 4%   -9.40%  (p=0.015 n=15+13)
IntersectionWithoutValidation/n=100-12                         61.6µs ±37%    58.0µs ±16%     ~     (p=0.170 n=13+15)
IntersectionWithoutValidation/n=1000-12                         271µs ±15%     273µs ±12%     ~     (p=0.928 n=13+15)
IntersectionWithoutValidation/n=10000-12                       2.66ms ±21%    2.79ms ±17%     ~     (p=0.201 n=14+15)
NoOp/n=10-12                                                   4.06µs ±13%    3.95µs ±16%     ~     (p=0.373 n=15+15)
NoOp/n=100-12                                                  10.8µs ±13%    10.7µs ±20%     ~     (p=0.880 n=13+13)
NoOp/n=1000-12                                                 73.9µs ±13%    77.2µs ±21%     ~     (p=0.387 n=13+15)
NoOp/n=10000-12                                                1.08ms ±46%    0.99ms ±19%     ~     (p=0.486 n=15+15)
pkg:github.com/peterstace/simplefeatures/internal/perf goos:linux goarch:amd64
LineStringIsSimpleCircle/n=10-12                               2.84µs ±92%    2.13µs ±24%  -24.96%  (p=0.023 n=15+15)
LineStringIsSimpleCircle/n=100-12                              33.2µs ±46%    29.1µs ± 8%  -12.61%  (p=0.037 n=15+15)
LineStringIsSimpleCircle/n=1000-12                              402µs ±11%     391µs ±23%     ~     (p=0.363 n=13+15)
LineStringIsSimpleCircle/n=10000-12                            5.64ms ±36%    5.33ms ±10%     ~     (p=0.496 n=13+15)
LineStringIsSimpleZigZag/10-12                                 2.00µs ±23%    1.81µs ±13%   -9.46%  (p=0.005 n=14+14)
LineStringIsSimpleZigZag/100-12                                36.3µs ±49%    30.6µs ±23%     ~     (p=0.187 n=15+15)
LineStringIsSimpleZigZag/1000-12                                395µs ±33%     350µs ±10%  -11.30%  (p=0.002 n=13+13)
LineStringIsSimpleZigZag/10000-12                              6.01ms ±36%    5.18ms ± 9%  -13.69%  (p=0.033 n=15+14)
…